cleaned compiler

+++ noumena/compiler.py
@@ -10,12 +10,16 @@ libnum = 0
 
 
 def bind_kernel(name):
-    # via ctypes interface
+
+    # attempt to load kernel
     try:
-        lib = CDLL('common/lib/' + name, RTLD_LOCAL)
+        lib = CDLL("tmp/" + name, RTLD_LOCAL)
+        os.system("rm tmp/" + name)
     except:
-        print "kernel compilation error.  exiting."
+        print "kernel not found.  exiting."
         exit()
+
+    # extract function
     kernel = lib.__device_stub_kernel_fb
     kernel.restype = None
     kernel.argtypes = [ c_void_p, c_ulong, c_void_p, c_int, c_float, c_float, c_float, c_float ]
@@ -24,37 +28,49 @@ def bind_kernel(name):
 
 
 class Compiler(threading.Thread):
+    ''' A Compiler object if responsible for asynchronously calling nvcc.
+        The compilation can be restarted by a call to update. '''
 
     def __init__(self, data, callback):
+
         self.data, self.callback = data, callback
-        threading.Thread.__init__(self)
+
+        # init update_vars
         self.update_vars = {}
         self.update_vars.update(data)
 
+        # init thread
+        threading.Thread.__init__(self)
+
+
     def update(self, new_vars):
+
+        # set update info
         self.update_vars.update(new_vars)
         self.do_update = True
 
+
     def render_file(self, name):
+
         # open file & read contents
         file = open("aer/" + name + ".ecu")
         contents = file.read()
         file.close()
 
-        if(name == "kernel"):
-            par_name_str = ""
+        # bind PAR_NAMES
+        par_name_str = ""
 
-            for i in xrange(len(self.data["par_names"])):
-                par_name_str += "#define " + self.data["par_names"][i] + " par[" + str(i) + "]\n"
+        for i in xrange(len(self.data["par_names"])):
+            par_name_str += "#define %s par[%d]\n" % (self.data["par_names"][i], i)
 
-            contents = re.compile('\%PAR_NAMES\%').sub(par_name_str, contents)
+        contents = re.compile('\%PAR_NAMES\%').sub(par_name_str, contents)
 
         # replace variables
         for key in self.update_vars:
-            contents = re.compile('\%' + key + '\%').sub(str(self.data[key]), contents)
+            contents = re.compile("\%" + key + "\%").sub(str(self.data[key]), contents)
 
         # write file contents
-        file = open("aer/__" + name + ".cu", 'w')
+        file = open("aer/__%s.cu" % name, 'w')
         file.write(contents)
         file.close()
 
@@ -63,20 +79,23 @@ class Compiler(threading.Thread):
 
         global libnum
 
+        # begin update loop
         self.do_update = True
-
         while(self.do_update):
             self.do_update = False
+
+            # render files
             self.render_file("seed")
             self.render_file("kernel")
 
-            name = "kernel" + str(libnum) + ".so"
-
+            # get name
+            name = "kernel%d.so" % libnum
             libnum += 1
 
-            os.system("rm common/lib/" + name)
-            os.system("/usr/local/cuda/bin/nvcc -Xcompiler -fPIC -o common/lib/%s --shared  aer/__kernel.cu" % name)
-            os.system("rm __kernel.linkinfo")
+            # compile
+            os.system("/usr/local/cuda/bin/nvcc -Xcompiler -fPIC -o tmp/%s --shared  aer/__kernel.cu" % name)
+            if(os.path.exists("__kernel.linkinfo")) : os.system("rm __kernel.linkinfo")
 
+        # execute callback
         self.callback(name)
